user info update endpoints merged
